refcat: common 책임 나누기
